point set refactoring
